separate props analysis to separate instance - `PropsAnalyzer`
